rustc: Extract native function generation into trans::native
